requirements.txt
Django==4.0.2
djangorestframework==3.13.1
psycopg2-binary==2.9.3
celery==5.2.3
redis==4.1.3
django-filter==21.1
django-taggit==2.1.0